fix(design-system): isolate design system package, and upgrade storybook
Isolated Design System apps. Upgraded Storybook, and also updated dependencies to use official
releases from the components and svg libraries.

BREAKING CHANGE: design-system apps is an isolated project, so this means, it will not be hoisted
into the yarn workspace, when you run `yarn install` in the root workspace. Any dependencies will be
free to be added without dependencies issues in the workspace.

fix #156, fix #194